Development

+++ campus/views.py
@@ -1,10 +1,14 @@
 from django.shortcuts import render
-from .models import *
+from datetime import datetime
+from reservas.models import Reserva
 from .forms import *
 
 
 def view_salas(request):
     try:
+        if not request.user.is_authenticated:
+            return render(request, 'permission_error.html')
+
         if request.user.tipo_usuario == 'CoordenadorEnsino':
             data = {}
             data['form'] = SalaForm(None)
@@ -33,30 +37,40 @@ def view_salas(request):
             data['form'] = SalaForm(None)
             return render(request, 'campus/salas/salas.html', data)
         else:
-            return render(request, 'campus/salas/permission_error.html')
+            return render(request, 'permission_error.html')
     except:
         data = {'mensagem': "Ocorreu um erro interno!" }
         return render(request, 'campus/salas/error.html', data)
 
 
-# Faltam as verificações quanto as reservas
 def delete_salas(request, id_sala, id_predio):
     try:
+        if not request.user.is_authenticated:
+            return render(request, 'permission_error.html')
+
         if request.user.tipo_usuario == 'CoordenadorEnsino':
-            data = {'mensagem': "Sala " + str(id_sala) + " removida com sucesso!"}
             predio = Predio.objects.get(id=id_predio)
-            predio.sala_set.get(pk=id_sala).delete()
-            return render(request, 'campus/salas/cadastro_sucesso.html', data)
+            sala = predio.sala_set.get(pk=id_sala)
+            reservas = Reserva.objects.filter(sala=sala, dataFim__gte=datetime.date.today().strftime("%Y-%m-%d"))
+            if len(reservas) > 0:
+                data = {'mensagem': "Não foi possível excluir a sala! Há reservas em aberto."}
+                return render(request, 'campus/salas/error.html', data)
+            else:
+                sala.delete()
+                data = {'mensagem': "Sala [" + str(sala) + "] removida com sucesso!"}
+                return render(request, 'campus/salas/cadastro_sucesso.html', data)
         else:
-            return render(request, 'campus/salas/permission_error.html')
+            return render(request, 'permission_error.html')
     except:
-        data = {'mensagem': "Não foi possível excluir a sala!"}
+        data = {'mensagem': "Ocorreu um erro interno!"}
         return render(request, 'campus/salas/error.html', data)
 
 
-# Faltam as verificações quanto as reservas
 def update_salas(request, id_sala, id_predio):
     try:
+        if not request.user.is_authenticated:
+            return render(request, 'permission_error.html')
+
         if request.user.tipo_usuario == 'CoordenadorEnsino':
             data = {}
             data['predios'] = Predio.objects.all()
@@ -69,29 +83,39 @@ def update_salas(request, id_sala, id_predio):
 
             form = SalaForm(request.POST or None, instance=sala)
             if form.is_valid():
-                sala = form.instance
-                sala.predio = data['predio']
-                sala.save()
-                data = {'mensagem': "Sala atualizada com sucesso!"}
-                return render(request, 'campus/salas/cadastro_sucesso.html', data)
+                reservas = Reserva.objects.filter(sala=sala, dataFim__gte=datetime.date.today().strftime("%Y-%m-%d"))
+                if len(reservas) == 0:
+                    form.instance.predio = data['predio']
+                    form.save()
+                    data = {'mensagem': "Sala atualizada com sucesso!"}
+                    return render(request, 'campus/salas/cadastro_sucesso.html', data)
+                else:
+                    data = {'mensagem': "Não foi possível atualizar a sala! Há reservas ativas!"}
+                    return render(request, 'campus/salas/error.html', data)
 
             data['form'] = form
             return render(request, 'campus/salas/update_salas.html', data)
         else:
-            return render(request, 'campus/salas/permission_error.html')
+            return render(request, 'permission_error.html')
     except:
-        data = {'mensagem': "Não foi possível atualizar a sala!"}
+        data = {'mensagem': "Ocorreu um erro interno!"}
         return render(request, 'campus/salas/error.html', data)
 
 
 def ver_sala(request, id_sala):
     data = {}
-    sala = Sala.objects.get(id=id_sala)
-    if sala:
-        data['sala'] = sala
-        data['equipamentos'] = sala.equipamentos.all()
-        return render(request, 'campus/salas/view_sala.html', data)
-    else:
+    if not request.user.is_authenticated:
+        return render(request, 'permission_error.html')
+    try:
+        sala = Sala.objects.get(id=id_sala)
+        if sala:
+            data['sala'] = sala
+            data['equipamentos'] = sala.equipamentos.all()
+            return render(request, 'campus/salas/view_sala.html', data)
+        else:
+            data = {'mensagem': "Não foi possível localizar a sala!"}
+            return render(request, 'campus/salas/error.html', data)
+    except:
         data = {'mensagem': "Não foi possível localizar a sala!"}
         return render(request, 'campus/salas/error.html', data)
 
@@ -99,6 +123,9 @@ def ver_sala(request, id_sala):
 
 def view_predios(request):
     try:
+        if not request.user.is_authenticated:
+            return render(request, 'permission_error.html')
+
         if request.user.tipo_usuario == 'CoordenadorEnsino':
             data = {}
             data['form'] = PredioForm(None)
@@ -127,28 +154,47 @@ def view_predios(request):
             data['form'] = PredioForm(None)
             return render(request, 'campus/predios/predios.html', data)
         else:
-            return render(request, 'campus/predios/permission_error.html')
+            return render(request, 'permission_error.html')
     except:
         data = {'mensagem': "Ocorreu um erro interno!" }
         return render(request, 'campus/predios/error.html', data)
 
-# Faltam as verificações quanto as reservas
+
 def delete_predios(request, id_predio, id_campus):
     try:
+        if not request.user.is_authenticated:
+            return render(request, 'permission_error.html')
+
         if request.user.tipo_usuario == 'CoordenadorEnsino':
-            data = {'mensagem': "Predio " + str(id_predio) + " removido com sucesso!"}
             campus = Campus.objects.get(id=id_campus)
-            campus.predio_set.get(pk=id_predio).delete()
-            return render(request, 'campus/predios/cadastro_sucesso.html', data)
+            predio = campus.predio_set.get(pk=id_predio)
+            salas = predio.sala_set.all()
+            removed = True
+            for sala in salas:
+                reservas = Reserva.objects.filter(sala=sala, dataFim__gte=datetime.date.today().strftime("%Y-%m-%d"))
+                if len(reservas) > 0:
+                    removed = False
+                    break
+
+            if removed:
+                predio.delete()
+                data = {'mensagem': "Predio " + str(id_predio) + " removido com sucesso!"}
+                return render(request, 'campus/predios/cadastro_sucesso.html', data)
+            else:
+                data = {'mensagem': "Não foi possível excluir o prédio! Há salas com reservas pendentes!"}
+                return render(request, 'campus/predios/error.html', data)
         else:
-            return render(request, 'campus/predios/permission_error.html')
+            return render(request, 'permission_error.html')
     except:
-        data = {'mensagem': "Não foi possível excluir o prédio!"}
+        data = {'mensagem': "Ocorreu um erro interno!"}
         return render(request, 'campus/predios/error.html', data)
 
-# Faltam as verificações quanto as reservas
+
 def update_predios(request, id_predio, id_campus):
     try:
+        if not request.user.is_authenticated:
+            return render(request, 'permission_error.html')
+
         if request.user.tipo_usuario == 'CoordenadorEnsino':
             data = {}
             data['campi'] = Campus.objects.all()
@@ -161,36 +207,58 @@ def update_predios(request, id_predio, id_campus):
 
             form = PredioForm(request.POST or None, instance=predio)
             if form.is_valid():
-                predio = form.instance
-                predio.campus = data['campus']
-                predio.save()
-                data = {'mensagem': "Prédio atualizado com sucesso!"}
-                return render(request, 'campus/predios/cadastro_sucesso.html', data)
+                salas = predio.sala_set.all()
+                update = True
+                for sala in salas:
+                    reservas = Reserva.objects.filter(sala=sala, dataFim__gte=datetime.date.today().strftime("%Y-%m-%d"))
+                    if len(reservas) > 0:
+                        update = False
+                        break
+
+                if update:
+                    predio = form.instance
+                    predio.campus = data['campus']
+                    predio.save()
+                    data = {'mensagem': "Prédio atualizado com sucesso!"}
+                    return render(request, 'campus/predios/cadastro_sucesso.html', data)
+                else:
+                    data = {'mensagem': "Não foi possível atualizar o prédio! Há salas com reservas pendentes!"}
+                    return render(request, 'campus/predios/error.html', data)
 
             data['form'] = form
             return render(request, 'campus/predios/update_predios.html', data)
         else:
-            return render(request, 'campus/predios/permission_error.html')
+            return render(request, 'permission_error.html')
     except:
         data = {'mensagem': "Não foi possível atualizar o prédio!"}
         return render(request, 'campus/predios/error.html', data)
 
+
 def ver_predio(request, id_predio):
     data = {}
-    predio = Predio.objects.get(id=id_predio)
-    if predio:
-        data['predio'] = predio
-        return render(request, 'campus/predios/view_predio.html', data)
-    else:
+    if not request.user.is_authenticated:
+        return render(request, 'permission_error.html')
+
+    try:
+        predio = Predio.objects.get(id=id_predio)
+        if predio:
+            data['predio'] = predio
+            return render(request, 'campus/predios/view_predio.html', data)
+        else:
+            data = {'mensagem': "Não foi possível localizar o prédio!"}
+            return render(request, 'campus/predios/error.html', data)
+    except:
         data = {'mensagem': "Não foi possível localizar o prédio!"}
         return render(request, 'campus/predios/error.html', data)
-
 
 
 ################# Campus ######################
 
 def view_campus(request):
     try:
+        if not request.user.is_authenticated:
+            return render(request, 'permission_error.html')
+
         if request.user.tipo_usuario == 'CoordenadorEnsino':
             data = {}
             campus = Campus.objects.all()
@@ -210,23 +278,34 @@ def view_campus(request):
             data['form'] = CampusForm(None)
             return render(request, 'campus/campus/campus.html', data)
         else:
-            return render(request, 'campus/campus/permission_error.html')
+            return render(request, 'permission_error.html')
     except:
         data = {'mensagem': "Ocorreu um erro interno!" }
         return render(request, 'campus/campus/error.html', data)
 
+
 def ver_campus(request, id_campus):
     data = {}
-    campus = Campus.objects.get(id=id_campus)
-    if campus:
-        data['campus'] = campus
-        return render(request, 'campus/campus/view_campus.html', data)
-    else:
+    if not request.user.is_authenticated:
+        return render(request, 'permission_error.html')
+    try:
+        campus = Campus.objects.get(id=id_campus)
+        if campus:
+            data['campus'] = campus
+            return render(request, 'campus/campus/view_campus.html', data)
+        else:
+            data = {'mensagem': "Não foi possível localizar o campus!"}
+            return render(request, 'campus/campus/error.html', data)
+    except:
         data = {'mensagem': "Não foi possível localizar o campus!"}
         return render(request, 'campus/campus/error.html', data)
 
+
 def update_campus(request, id_campus):
     try:
+        if not request.user.is_authenticated:
+            return render(request, 'permission_error.html')
+
         if request.user.tipo_usuario == 'CoordenadorEnsino':
             data = {}
             data['campi'] = Campus.objects.all()
@@ -236,11 +315,23 @@ def update_campus(request, id_campus):
             campus = Campus.objects.get(id=id_campus)
 
             form = CampusForm(request.POST or None, instance=campus)
+            update = True
             if form.is_valid():
-                campus = form.instance
-                campus.save()
-                data = {'mensagem': "Campus atualizado com sucesso!"}
-                return render(request, 'campus/campus/cadastro_sucesso.html', data)
+                # Se horário de funcionamento do campus foi alterado
+                if form.instance.horaInicio != data['campus'].horaInicio or form.instance.horaFim != data['campus'].horaFim:
+                    reservas = Reserva.objects.filter(dataFim__gte=datetime.date.today().strftime("%Y-%m-%d"))
+                    print("Horarios diferentes")
+                    if len(reservas) > 0:
+                        update = False
+
+                if update:
+                    campus = form.instance
+                    campus.save()
+                    data = {'mensagem': "Campus atualizado com sucesso!"}
+                    return render(request, 'campus/campus/cadastro_sucesso.html', data)
+                else:
+                    data = {'mensagem': "Horário de funcionamento não pode ser atualizado, há reservas em aberto!"}
+                    return render(request, 'campus/campus/error.html', data)
 
             data['form'] = form
             return render(request, 'campus/campus/update_campus.html', data)
@@ -254,6 +345,9 @@ def update_campus(request, id_campus):
 
 def view_equips(request):
     try:
+        if not request.user.is_authenticated:
+            return render(request, 'permission_error.html')
+
         if request.user.tipo_usuario == 'CoordenadorEnsino':
             data = {}
             form = EquipForm(request.POST or None)
@@ -270,25 +364,34 @@ def view_equips(request):
             data['form'] = form
             return render(request, 'campus/equipamentos/equipamentos.html', data)
         else:
-            return render(request, 'campus/equipamentos/permission_error.html')
+            return render(request, 'permission_error.html')
     except:
-        data = {'mensagem': "Ocorreu um erro interno!" }
+        data = {'mensagem': "Ocorreu um erro interno!"}
         return render(request, 'campus/equipamentos/error.html', data)
 
 
 def ver_equip(request, id_equip):
     data = {}
-    equip = Equipamento.objects.get(id=id_equip)
-    if equip:
-        data['equip'] = equip
-        return render(request, 'campus/equipamentos/view_equipamento.html', data)
-    else:
+    if not request.user.is_authenticated:
+        return render(request, 'permission_error.html')
+    try:
+        equip = Equipamento.objects.get(id=id_equip)
+        if equip:
+            data['equip'] = equip
+            return render(request, 'campus/equipamentos/view_equipamento.html', data)
+        else:
+            data = {'mensagem': "Não foi possível localizar o equipamento!"}
+            return render(request, 'campus/equipamentos/error.html', data)
+    except:
         data = {'mensagem': "Não foi possível localizar o equipamento!"}
         return render(request, 'campus/equipamentos/error.html', data)
 
 
 def update_equip(request, id_equip):
     try:
+        if not request.user.is_authenticated:
+            return render(request, 'permission_error.html')
+
         if request.user.tipo_usuario == 'CoordenadorEnsino':
             data = {}
             data['equipamentos'] = Equipamento.objects.all()
@@ -297,14 +400,19 @@ def update_equip(request, id_equip):
             form = EquipForm(request.POST or None, instance=data['equipamento'])
 
             if form.is_valid():
-                form.save()
-                data = {'mensagem': "Equipamento atualizado com sucesso!"}
-                return render(request, 'campus/equipamentos/cadastro_sucesso.html', data)
+                salas = data['equipamento'].sala_set.all()
+                if len(salas) == 0:
+                    form.save()
+                    data = {'mensagem': "Equipamento atualizado com sucesso!"}
+                    return render(request, 'campus/equipamentos/cadastro_sucesso.html', data)
+                else:
+                    data = {'mensagem': "Não é possível atualizar equipamentos que estão em uso em salas!"}
+                    return render(request, 'campus/equipamentos/error.html', data)
 
             data['form'] = form
             return render(request, 'campus/equipamentos/update_equipamento.html', data)
         else:
-            return render(request, 'campus/equipamentos/permission_error.html')
+            return render(request, 'permission_error.html')
     except:
         data = {'mensagem': "Não foi possível atualizar o equipamento!"}
         return render(request, 'campus/equipamentos/error.html', data)
@@ -312,13 +420,21 @@ def update_equip(request, id_equip):
 
 def delete_equip(request, id_equip):
     try:
+        if not request.user.is_authenticated:
+            return render(request, 'permission_error.html')
+
         if request.user.tipo_usuario == 'CoordenadorEnsino':
-            data = {'mensagem': "Equipamento " + str(id_equip) + " removido com sucesso!"}
             equip = Equipamento.objects.get(id=id_equip)
-            equip.delete()
-            return render(request, 'campus/equipamentos/cadastro_sucesso.html', data)
+            salas = equip.sala_set.all()
+            if len(salas) == 0:
+                equip.delete()
+                data = {'mensagem': "Equipamento " + str(id_equip) + " removido com sucesso!"}
+                return render(request, 'campus/equipamentos/cadastro_sucesso.html', data)
+            else:
+                data = {'mensagem': "Não é possível excluir equipamentos que estão em uso em salas!"}
+                return render(request, 'campus/equipamentos/error.html', data)
         else:
-            return render(request, 'campus/equipamentos/permission_error.html')
+            return render(request, 'permission_error.html')
     except:
         data = {'mensagem': "Não foi possível excluir o equipamento!"}
         return render(request, 'campus/equipamentos/error.html', data)
